// app/(employees)/shifts/types.ts
  export interface JobCardProps {
    jobId: string;
    date?: string;
    time?: string;
    location?: string;
    employee?: string;
    position?: string;
    hourlyRate?: number;
    hours?: number;
    totalAmount?: number;
    backgroundImage?: any;
  }